requirements.txt
django==4.0.4
Pillow==9.1.0
signals==0.0.2
django-taggit==3.0.0
